Move tests for T2 chassis frontend to separate files
